+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Impl `TryFrom<elephantry::Interval>` for `chrono::Duration`;
 - Inspect constraints;
 - New `FromText`/`ToText` to easily convert type from/to text;
+- New `inspect::Relation::kind` field;
+
+### Changed
+- `inspect::Relation::ty` is deprecated;
 
 ### Fixed
 - `u8` conversion;